backend/requirements.txt
Django==2.2.28
Pillow==9.4.0
pytz==2022.7.1
sqlparse==0.4.3
djangorestframework==3.13.1
djoser==2.1.0
